requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.133
botocore==1.34.133
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
python-dotenv==1.0.1
scikit-learn==1.5.0
scipy==1.14.0
tqdm==4.66.4
matplotlib==3.9.0